refactor: convert all entity IDs from Long to Integer
Update all JPA entity models, Spring Data repositories, DTOs, and associated service/controller logic to use Integer instead of Long for primary keys and foreign ID fields
Modify SQL schema scripts and Flyway migration files to use INT(11) instead of BIGINT(20) for auto-incrementing primary keys
Update the application datasource URL to connect to the gppremiumfabrica database instead of gppremium
Remove the unused MotoristaControllerTest class
Fix pom.xml: remove the unnecessary skip test flag and add the missing final newline
Add a port-killing command to README.md to simplify local development port cleanup
Clean up accidental macOS .DS_Store system files

+++ scripts/maquina_vulc.sql
@@ -63,7 +63,7 @@ DEFAULT CHARACTER SET = utf8mb4
 COLLATE = utf8mb4_0900_ai_ci;
 
 CREATE TABLE IF NOT EXISTS `gppremium`.`pneus_vulcanizados` (
-  `id` BIGINT(20) NOT NULL AUTO_INCREMENT,
+  `id` INT(11) NOT NULL AUTO_INCREMENT,
   `dt_create` DATETIME NOT NULL,
   `dt_delete` DATETIME NULL DEFAULT NULL,
   `dt_update` DATETIME NULL DEFAULT NULL,
@@ -91,7 +91,7 @@ COLLATE = utf8mb4_0900_ai_ci
 COMMENT = 'Tabela para controle de pneus vulcanizados por usuário e produção';
 
 CREATE TABLE IF NOT EXISTS `gppremium`.`rele` (
-  `id` BIGINT(20) NOT NULL AUTO_INCREMENT,
+  `id` INT(11) NOT NULL AUTO_INCREMENT,
   `dt_create` DATETIME NOT NULL,
   `dt_delete` DATETIME NULL DEFAULT NULL,
   `dt_update` DATETIME NULL DEFAULT NULL,
