+++ src/main/java/br/compneusgppremium/api/repository/SonoffRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 import java.util.List;
 
-public interface SonoffRepository extends JpaRepository<SonoffModel, Long> {
+public interface SonoffRepository extends JpaRepository<SonoffModel, Integer> {
     Optional<SonoffModel> findByIpAndDtDeleteIsNull(String ip);
 
     List<SonoffModel> findByCelularIdAndDtDeleteIsNullOrderByDtCreateDesc(String celularId);
